+++ src/SystemVariables.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Ergebnis\Environment;
 
-final class Production implements Variables
+final class SystemVariables implements Variables
 {
     public function has(string $name): bool
     {